Button Press Demo
These projects were used to demo the functionality of CC1310 boards. The motivation behind this demo was to allow people to hold a board and press a button to see how it was communicating with other boards. The button press would cause the held board to transmit, which would cause the receiving board to modify its pin output values, as specified by the user. This could be utilized to blink an LED or hookup the receiving board to a more complicated system. For demo purposes, receiving boards were connected to power relays which were able to activate and deactivate various appliances at the demo.